Move ServerStatus
